upcode 23

+++ app/Http/Controllers/CateGoryController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 class CategoryController extends Controller
@@ -13,16 +15,34 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
+//        dd($request);
+        $params = $request->all();
+//        dd($params);
+        $filter_type = 2;
+        if($request->has('filter_type'))  {
+            $filter_type = $params['filter_type'];
+        }
+
+//        dd($filter_type);
         //Cách 1: Lấy toàn bộ dữ liệu
         //$data = Category::all(); // SELECT * FROM Categorys
 
-        //Cách 2: Lấy dữ liệu mới nhất và phân trang - mỗi trang 10 bản ghi
-        $data = Category::latest()->paginate(10);
+        // if check admin
+        if (Auth::user()->role_id == 1) {
+            if ($filter_type == 1) {
+                $data = Category::withTrashed()->latest()->paginate(10);
+            } elseif ($filter_type == 2) {
+                $data = Category::latest()->paginate(10);
+            } elseif ($filter_type == 3){
+                $data = Category::onlyTrashed()->latest()->paginate(10);
+            }
 
-
-        return view('backend.category.index', ['data' => $data]);
+        } else {
+            $data = Category::latest()->paginate(10);
+        }
+        return view('backend.category.index', ['data' => $data, 'filter_type' => $filter_type]);
     }
 
     /**
@@ -194,14 +214,31 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $Category = Category::findOrFail($id);
-        // xóa ảnh cũ
-        @unlink(public_path($Category->image));
+
+        $checkExitsProduct = Product::where('category_id',$id)->first();
+
+        if($checkExitsProduct != null){
+            return response()->json([
+                'status' => false,
+                'msg' => 'Xóa thất bại vì có tồn tại sản phẩm trong danh mục này!!!'
+            ]);
+        }
 
         Category::destroy($id);
 
         return response()->json([
             'status' => true,
             'msg' => 'Xóa thành công'
+        ]);
+    }
+    public function restore($id)
+    {
+        $Category = Category::onlyTrashed()->findOrFail($id);
+        $Category->restore();
+
+        return response()->json([
+            'status' => true,
+            'msg' => 'Khôi phục thành công'
         ]);
     }
 }
