+++ resources/views/backend/category/index.blade.php
@@ -17,6 +17,15 @@
             <div class="col-md-12">
                 <div class="box">
                     <div class="box-header with-border">
+                        @if(\Auth::user()->role_id == 1)
+                            <div class="form-group" style="width: 150px;float: left;margin: 0">
+                                <select class="form-control" id="filter_type" name="filter_type">
+                                    <option {{ $filter_type == 1 ? 'selected' : '' }} value="1">Tất cả</option>
+                                    <option {{ $filter_type == 2 ? 'selected' : '' }} value="2">Đang Sử Dụng</option>
+                                    <option {{ $filter_type == 3 ? 'selected' : '' }} value="3">Đã Bị Xóa</option>
+                                </select>
+                            </div>
+                        @endif
                         <a href="{{ route('admin.category.create') }}" class="btn btn-primary pull-right"><i class="fa fa-plus" aria-hidden="true"></i></a>
                     </div>
                     <!-- /.box-header -->
@@ -54,9 +63,15 @@
                                 <td>
                                     {{ $item->position }}
                                 </td>
-                                <td>
+                                <td class="action" >
                                     <a href="{{ route('admin.category.edit', ['category' => $item->id]) }}"><span title="Chỉnh sửa" type="button" class="btn btn-flat btn-primary"><i class="fa fa-edit"></i></span></a>
-                                    <span data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                    @if($item->deleted_at == null)
+                                        <span  data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                        <span style="display:none;" data-id="{{ $item->id }}" title="Khôi phục" class="btn btn-flat btn-warning restoreItem"><i class="fa fa-refresh"></i></span>
+                                    @else
+                                        <span style="display:none;" data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                        <span  data-id="{{ $item->id }}" title="Khôi phục" class="btn btn-flat btn-warning restoreItem"><i class="fa fa-refresh"></i></span>
+                                    @endif
                                 </td>
                             </tr>
                             @endforeach
@@ -107,6 +122,49 @@
                         });
                     }
                 });
+            });
+            $('.restoreItem').click(function () {
+                var id = $(this).attr('data-id');
+                Swal.fire({
+                    title: 'Bạn có muốn khôi phục ?',
+                    text: "Dữ liệu khôi phục sẽ được nhìn thấy bởi tất cả các thành viên",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'OK'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $.ajax({
+                            url : '/admin/category/restore/'+id,
+                            type: 'POST',
+                            data: {},
+                            success: function (res) {
+                                if(res.status) {
+                                    Swal.fire(
+                                        'Thông báo !',
+                                        'Khôi phục thành công',
+                                        'success'
+                                    )
+                                    $('.restoreItem').remove();
+                                    $('.deleteItem').show();
+                                } else {
+                                    Swal.fire(
+                                        'Thông báo !',
+                                        'Có lỗi xảy ra',
+                                        'error'
+                                    )
+                                }
+                            },
+                            error: function (res) {
+                            }
+                        });
+                    }
+                });
+            });
+            $('#filter_type').change(function () {
+                var filter_type = $('#filter_type').val();
+                window.location.href = "{{ route('admin.category.index') }}?filter_type="+filter_type;
             });
         });
     </script>
